+++ src/Toml/Codec/BiMap/Conversion.hs
@@ -64,6 +64,7 @@ module Toml.Codec.BiMap.Conversion
       -- * 'Key's
     , _KeyText
     , _KeyString
+    , _KeyInt
 
       -- * General purpose
     , _Just
@@ -575,6 +576,17 @@ _KeyString :: TomlBiMap Key String
 _KeyString = BiMap
     { forward = Right . T.unpack . prettyKey
     , backward = textToKey . T.pack
+    }
+
+{- | Bidirectional converter between 'Key' and 'Int'. Usually used
+as an argument for 'Toml.Codec.Combinator.Map.tableIntMap'.
+
+@since 1.3.0.0
+-}
+_KeyInt :: TomlBiMap Key Int
+_KeyInt = BiMap
+    { forward = first (ArbitraryError . T.pack) . readEither . T.unpack . prettyKey
+    , backward = textToKey . tShow
     }
 
 textToKey :: Text -> Either TomlBiMapError Key